LayoutTests/imported/w3c/web-platform-tests/dom/nodes/Node-baseURI-expected.txt


PASS For elements belonging to document, baseURI should be document url 
FAIL For elements unassigned to document, baseURI should be document url assert_equals: expected (string) "http://localhost:8800/dom/nodes/Node-baseURI.html" but got (object) null
FAIL For elements belonging to document fragments, baseURI should be document url assert_equals: expected (string) "http://localhost:8800/dom/nodes/Node-baseURI.html" but got (object) null
PASS After inserting fragment into document, element baseURI should be document url 
